requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-RateLimiter==0.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
mysqlclient==2.2.0
PyJWT==2.8.0
python-dateutil==2.8.2
pymysql==1.1.0
cryptography==41.0.3
email-validator==2.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
setuptools==68.2.2
wheel==0.41.2
